refactor: move helper functions to own file
